CAMEL-14050: camel-main - Add logic for automatic RouteBuilder class detection ala camel-spring-boot has
