+++ components/camel-spring-boot/src/main/java/org/apache/camel/spring/boot/RoutesCollector.java
@@ -17,7 +17,6 @@
 package org.apache.camel.spring.boot;
 
 import java.io.FileNotFoundException;
-import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
@@ -30,6 +29,7 @@ import org.apache.camel.CamelContext;
 import org.apache.camel.RoutesBuilder;
 import org.apache.camel.StartupListener;
 import org.apache.camel.main.MainDurationEventNotifier;
+import org.apache.camel.main.MainRoutesCollector;
 import org.apache.camel.model.Model;
 import org.apache.camel.model.ModelHelper;
 import org.apache.camel.model.RoutesDefinition;
@@ -40,7 +40,6 @@ import org.apache.camel.spi.EventNotifier;
 import org.apache.camel.support.EventNotifierSupport;
 import org.apache.camel.support.OrderedComparator;
 import org.apache.camel.support.service.ServiceHelper;
-import org.apache.camel.util.ObjectHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.ApplicationContext;
@@ -49,7 +48,6 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.core.Ordered;
 import org.springframework.core.io.Resource;
-import org.springframework.util.AntPathMatcher;
 
 /**
  * Collects routes and rests from the various sources (like Spring application context beans registry or opinionated
@@ -89,68 +87,8 @@ public class RoutesCollector implements ApplicationListener<ContextRefreshedEven
                 && camelContext.getStatus().isStopped()) {
             LOG.debug("Post-processing CamelContext bean: {}", camelContext.getName());
 
-            final AntPathMatcher matcher = new AntPathMatcher();
-            final List<RoutesBuilder> routes = new ArrayList<>();
-            for (RoutesBuilder routesBuilder : applicationContext.getBeansOfType(RoutesBuilder.class, configurationProperties.isIncludeNonSingletons(), true).values()) {
-                // filter out abstract classes
-                boolean abs = Modifier.isAbstract(routesBuilder.getClass().getModifiers());
-                if (!abs) {
-                    String name = routesBuilder.getClass().getName();
-                    // make name as path so we can use ant path matcher
-                    name = name.replace('.', '/');
-
-                    String exclude = configurationProperties.getJavaRoutesExcludePattern();
-                    String include = configurationProperties.getJavaRoutesIncludePattern();
-
-                    boolean match = !"false".equals(include);
-                    // exclude take precedence over include
-                    if (match && ObjectHelper.isNotEmpty(exclude)) {
-                        // there may be multiple separated by comma
-                        String[] parts = exclude.split(",");
-                        for (String part : parts) {
-                            // must negate when excluding, and hence !
-                            match = !matcher.match(part, name);
-                            LOG.trace("Java RoutesBuilder: {} exclude filter: {} -> {}", name, part, match);
-                            if (!match) {
-                                break;
-                            }
-                        }
-                    }
-                    // special support for testing with @ExcludeRoutes annotation with camel-test-spring
-                    exclude = System.getProperty("CamelTestSpringExcludeRoutes");
-                    // exclude take precedence over include
-                    if (match && ObjectHelper.isNotEmpty(exclude)) {
-                        // this property is a comma separated list of FQN class names, so we need to make
-                        // name as path so we can use ant patch matcher
-                        exclude = exclude.replace('.', '/');
-                        // there may be multiple separated by comma
-                        String[] parts = exclude.split(",");
-                        for (String part : parts) {
-                            // must negate when excluding, and hence !
-                            match = !matcher.match(part, name);
-                            LOG.trace("Java RoutesBuilder: {} exclude filter: {} -> {}", name, part, match);
-                            if (!match) {
-                                break;
-                            }
-                        }
-                    }
-                    if (match && ObjectHelper.isNotEmpty(include)) {
-                        // there may be multiple separated by comma
-                        String[] parts = include.split(",");
-                        for (String part : parts) {
-                            match = matcher.match(part, name);
-                            LOG.trace("Java RoutesBuilder: {} include filter: {} -> {}", name, part, match);
-                            if (match) {
-                                break;
-                            }
-                        }
-                    }
-                    LOG.debug("Java RoutesBuilder: {} accepted by include/exclude filter: {}", name, match);
-                    if (match) {
-                        routes.add(routesBuilder);
-                    }
-                }
-            }
+            MainRoutesCollector collector = new MainRoutesCollector();
+            final List<RoutesBuilder> routes = collector.collectRoutes(camelContext, configurationProperties);
 
             // sort routes according to ordered
             routes.sort(OrderedComparator.get());
@@ -299,6 +237,7 @@ public class RoutesCollector implements ApplicationListener<ContextRefreshedEven
     }
 
     // Helpers
+
 
     private void loadXmlRoutes(ApplicationContext applicationContext, CamelContext camelContext, String directory) throws Exception {
         String[] parts = directory.split(",");
